Rename textarea to text_area and customize the macro
